Refactor has_many_extension
